% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoByGene.R
\name{annoByGene}
\alias{annoByGene}
\title{Annotate integration sites by genes and TSSs.}
\usage{
annoByGene(hits, ran_hits = NULL,
           mapTool = 'blast',
           organism = 'GRCh37',
           interval = 5000,
           range = c(-20000, 20000),
           outPath = getwd(),
           outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{hits}{GR object. This object made by \code{\link{makeExpSet}} function.}

\item{ran_hits}{GR object or list. This object is output of \code{\link{makeRanSet}} function.}

\item{mapTool}{Character. Function uses two types of object\cr from BLAST and BLAT.
Default is 'blast'. If you want to use BLAT result, use 'blat'.}

\item{organism}{Character. This function can run by two versions of organisms\cr
such as GRCh37, GRCh38 (Human). Default is 'GRCh37'.}

\item{interval}{Integer. This number means interval number\cr for
distribution analysis. Default is 5000.}

\item{range}{Integer array. The range of highlight region for analysis.\cr
Default range is \code{c(-20000, 20000)}.}

\item{outPath}{String. Plots are saved in this path. \cr Default value is R home directory.}

\item{outFileName}{Character. Attached ID to the result file name.}
}
\value{
Return a result list that is made up of insertion and distribution result tables\cr and GenomicRange object of Gene and TSS data.
}
\description{
Annotate vector integration sites by gene data.
}
\examples{
data(blast_obj); data(gene_exam_db); data(tss_exam_db)
saveRDS(gene_exam_db, 
        paste0(system.file("extdata", package = 'RIPAT'), 
               '/GRCh37_gene.rds'))
saveRDS(tss_exam_db, 
        paste0(system.file("extdata", package = 'RIPAT'),
               '/GRCh37_TSS.rds'))

blast_gene = annoByGene(hits = blast_obj, ran_hits = NULL, 
                        outFileName = 'blast_res')

}
