% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoByRepeat.R
\name{annoByRepeat}
\alias{annoByRepeat}
\title{Annotate integration sites by repeats and microsatellites.}
\usage{
annoByRepeat(hits, ran_hits = NULL, mapTool = 'blast',
             organism = 'GRCh37', interval = 5000, range = c(-20000, 20000), 
             includeUndecided = FALSE, outPath = getwd(),
             outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{hits}{a GR object. This object made by \code{makeInputObj} function.}

\item{ran_hits}{a GR object or list. This object is output of \code{ranSetGenerator} function.}

\item{mapTool}{a single character. Function serves two types of object
such as outputs from BLAST and BLAT.
Default is 'blast'. If you want to use BLAT result, use 'blat'.}

\item{organism}{a single character. This function can run by two versions of organisms
such as GRCh37, GRCh38 (Human). Default is 'GRCh37'.}

\item{interval}{an integer vector. This number means interval number for
distribution analysis. Default is 5000.}

\item{range}{an integer array. The range of highlight region for analysis.
Default range is c(-20000, 20000).}

\item{includeUndecided}{TRUE or FALSE. If user want to use undecided hits in analysis,
enter TRUE. Default is FALSE.}

\item{outPath}{an string vector. Plots are saved in this path. Default value is R home directory.}

\item{outFileName}{a character vector. Attached ID to the result file name.}
}
\value{
Return a result list that is made up of insertion and distribution result tables
        and GenomicRange object of Rpeat and microsatellite data.
}
\description{
Annotate vector integration sites by repeat and microsatellite data.
}
\examples{
data(blast_obj); data(repeat_exam_db); data(micro_exam_db)
saveRDS(repeat_exam_db, paste0(system.file("extdata", package = 'RIPAT'), '/GRCh37_repeat.rds'))
saveRDS(micro_exam_db, paste0(system.file("extdata", package = 'RIPAT'), '/GRCh37_microsat.rds'))

blast_repeat = annoByRepeat(hits = blast_obj, ran_hits = NULL, outFileName = 'blast_res')

}
