% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_gene.R
\docType{data}
\name{blast_gene}
\alias{blast_gene}
\title{Integration site annotation by gene data}
\format{List object}
\usage{
data(blast_gene)
}
\description{
\preformatted{
This data is generated by annoByGene function on random set analysis mode. 
Plus, it is used in example code of \code{drawingKaryo} and \code{makeDocument}.
}
}
\examples{
data(blast_gene)

}
\keyword{datasets}
