% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_obj.R
\docType{data}
\name{blast_obj}
\alias{blast_obj}
\title{Retroviral vector integration site object}
\format{GRange object}
\usage{
data(blast_obj)
}
\description{
Data used in example code of integration site annotation functions
}
\examples{
data(blast_obj)

}
\keyword{datasets}
