% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCases.R
\name{compareCases}
\alias{compareCases}
\title{Annotate integration sites by genes and TSSs.}
\usage{
annoByGene(res1, res2,
           outPath = getwd(),
           outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{res1}{GR object. This object made by \code{\link{makeExpSet}} function.}

\item{res2}{GR object or list. This object is output of \code{\link{makeRanSet}} function.}

\item{outPath}{String. Plots are saved in this path. \cr Default value is R home directory.}

\item{outFileName}{Character. Attached ID to the result file name.}
}
\value{
Return a result list that is made up of insertion and distribution result tables\cr and GenomicRange object of Gene and TSS data.
}
\description{
Annotate vector integration sites by gene data.
}
\examples{
data(blast_res); data(blast_res2)

blast_gene = compareCases(res1 = blast_res, res2 = blast_res2,
                          outFileName = 'blast_res')

}
