% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_exam_db.R
\docType{data}
\name{cpg_exam_db}
\alias{cpg_exam_db}
\title{Data file for annoByCpG vignette}
\format{Data table object}
\usage{
data(cpg_exam_db)
}
\description{
Data used in example code of \code{annoByCpG}
}
\examples{
data(cpg_exam_db)

}
\keyword{datasets}
