% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawingKaryo.R
\name{drawingKaryo}
\alias{drawingKaryo}
\title{Draw the karyogram plot.}
\usage{
drawingKaryo(hits, feature,
             organism = 'GRCh37',
             outPath = getwd(), 
             outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{hits}{a GR object. This object made from \code{\link{makeExpSet}} function.}

\item{feature}{a GR object. This object made from annotation function.}

\item{organism}{a character vector. This function serves 2 versions of organisms\cr
such as GRCh37, GRCh38. Default is 'GRCh37'.}

\item{outPath}{a string vector. Type path to save a plot.}

\item{outFileName}{a character vector. This value used when saving the idegoram image file.}
}
\value{
Return the ideogram plot and object.
}
\description{
Draw a karyogram plot and show integration site.
}
\examples{
data(blast_obj); data(blast_gene)
drawingKaryo(hits = blast_obj, feature = blast_gene$Gene_data, outFileName = 'blast_res')

}
