% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exam_db.R
\docType{data}
\name{gene_exam_db}
\alias{gene_exam_db}
\title{Data file for annoByGene vignette}
\format{Data table object}
\usage{
data(gene_exam_db)
}
\description{
\preformatted{
Data used in example code of \code{annoByGene}.
}
}
\examples{
data(gene_exam_db)

}
\keyword{datasets}
