% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Make data files for RIPAT.}
\usage{
makeData(organism = 'GRCh37', dataType = 'cpg')
}
\arguments{
\item{organism}{a single character. Two versions of organism such as GRCh37, GRCh38 (Human).\cr
Default is 'GRCh37'.}

\item{dataType}{a single character. Data type what user needs (cpg, repeat and variant).\cr
Default is 'cpg'.}
}
\description{
Download datafiles for running RIPAT.
}
\examples{
makeData(organism = 'GRCh37')

}
