% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExpSet.R
\name{makeExpSet}
\alias{makeExpSet}
\title{Make the retroviral vector integration site object.}
\usage{
makeExpSet(inFile,
           mapTool = 'blast', 
           vectorPos = 'front',
           outPath = getwd(), 
           outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{inFile}{a string vector. The path of a local alignment result file.\cr
File do not include any header and comment.}

\item{mapTool}{a character vector. Function serves two types of file\cr
such as outputs from BLAST and BLAT. Default is 'blast'.\cr
If you want to use BLAT result, use 'blat'.}

\item{vectorPos}{a character vector. Sets the position of vector on sequences.\cr
Default value is 'front'. If the vector is located at\cr
the behind of sequence, you can change it to 'behind'.\cr}

\item{outPath}{a string vector. Directory path of tab-deliminated hit files\cr
generated by this function.}

\item{outFileName}{a character vector. Attached character to the result file name.}
}
\value{
Return two types of outputs. Text file and R object.\cr
        Available hit data from input is written in text file and\cr
        generated as a list of GenomicRange(GR) format object.
}
\description{
Make an input object for annotation functions.
}
\examples{
blast_obj = makeExpSet(inFile = paste0(.libPaths()[1],
                                       '/RIPAT/scripts/A5_15856M_BLASTn.txt'))

}
