% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInputObj.R
\name{makeInputObj}
\alias{makeInputObj}
\title{Make an retroviral vector integration site object}
\usage{
makeInputObj(inFile, mapTool = 'blast', 
             vectorPos = 'front', outPath = getwd(), 
             outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{inFile}{a string vector. Location of a local alignment result file. When you use BLAT result file (*.psl),
do not include any header and comment.}

\item{mapTool}{a character vector. Function serves two types of file such as outputs from BLAST and BLAT.
Default is 'blast'. If you want to use file from BLAT, use 'blat'.}

\item{vectorPos}{a character vector. Sets the position of the vector. Default value is 'front'. If the vector is located at the back, you can change it to 'behind'.}

\item{outPath}{a string vector. Directory path of tab-deliminated hit files.}

\item{outFileName}{a character vector. Attached ID to the result file name.}
}
\value{
Return integration site objects in two types by a list. Data table format object and GenomicRange(GR) format object.
}
\description{
\preformatted{
Annotation funcions in this package need specific format of inputs.
So, this function converts raw local alignment results to objects with the avaliable format. 
User can use BLAST and BLAT output format specifically.
}
}
\examples{
#' 
blast_obj = makeInputObj(inFile = paste0(.libPaths()[1], '/RIPAT/scripts/A5_15856M_BLASTn.txt'))

}
