% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInputObj2.R
\name{makeInputObj2}
\alias{makeInputObj2}
\title{Make an retroviral vector integration site object}
\usage{
makeInputObj2(inDir, id, mapTool = 'blast', 
              vectorPos = 'front', outPath = getwd(), 
              outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{inDir}{a string vector. Location of a directory that saves a local alignment result files.
All alignment result files do not include any header and comment line.}

\item{id}{a character vector. For this function, users should write the file name with specific words
that can appoint the alignment files exclusively.}

\item{mapTool}{a character vector. Function serves two types of file such as outputs from BLAST and BLAT.
Default is 'blast'. If you want to use file from BLAT, use 'blat'.}

\item{vectorPos}{a character vector. Sets the position of the vector. Default value is 'front'. If the vector is located at the back, you can change it to 'behind'.}

\item{outPath}{a string vector. Directory path of tab-deliminated hit files.}

\item{outFileName}{a character vector. Attached ID to the result file name.}
}
\value{
Return integration site objects in two types by a list. Data table format object and GenomicRange(GR) format object.
}
\description{
\preformatted{
Annotation funcions in this package need specific format of inputs.
So, this function converts raw local alignment results to objects with the avaliable format. 
Specifically, User can use BLAST and BLAT output format.
}
}
\examples{

blast_obj = makeInputObj2(inDir = paste0(.libPath()[1], '/RIPAT/scripts'), id = 'BLASTn')

 
}
