% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_exam_db.R
\docType{data}
\name{micro_exam_db}
\alias{micro_exam_db}
\title{Data file for annoByRepeat vignette}
\format{Data table object}
\usage{
data(micro_exam_db)
}
\description{
\preformatted{
Data used in example code of \code{annoByRepeat}.
}
}
\examples{
data(micro_exam_db)

}
\keyword{datasets}
