% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranSetGenerator.R
\name{ranSetGenerator}
\alias{ranSetGenerator}
\title{Random data of viral vector integration site}
\usage{
ranSetGenerator(organism = 'GRCh37', randomSize = if(doRandom){10000}else{NULL}, outPath = getwd(),
              outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{organism}{a single character. This function can run by two versions of organisms
such as GRCh37, GRCh38 (Human). Default is 'GRCh37'.}

\item{randomSize}{an integer vector. A random set size. Default is 10000.}

\item{outPath}{an string vector. Output file is saved in this path. Default value is R home directory.}

\item{outFileName}{a character vector. Attached ID to the result file name.}
}
\value{
Return a list of integration site positions on chromosome as an R object and text file.
}
\description{
Generate viral vector integration site random data
}
\examples{
ranSetGenerator(organism = 'GRCh37')

}
