% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_exam_db.R
\docType{data}
\name{var_exam_db}
\alias{var_exam_db}
\title{Data file for annoByVar vignette}
\format{Data table object}
\usage{
data(var_exam_db)
}
\description{
\preformatted{
Data used in example code of \code{annoByVar}.
}
}
\examples{
data(var_exam_db)

}
\keyword{datasets}
