% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCases.R
\name{compareCases}
\alias{compareCases}
\title{Annotate integration sites by genes and TSSs.}
\usage{
compareCases(res_list, excelOut = TRUE,
             outPath = getwd(),
             outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{res_list}{List. This is the list of result made by \code{\link{annoByGene}}, \code{\link{annoByCpG}},
\code{\link{annoByRepeat}}, \code{\link{annoByVar}} and \code{\link{annoByCGC}} function.}

\item{excelOut}{TRUE or FALSE. If user want to make excel file, enter TRUE.\cr
Default is TRUE.}

\item{outPath}{String. Outputs are saved in this path. \cr Default value is R home directory.}

\item{outFileName}{Character. Attached ID to the result file name.}
}
\value{
Return a result list that is made up of distribution plot and table.
}
\description{
Annotate vector integration sites by gene data.
}
\examples{
data(blast_gene)
in_list = list(blast_gene, blast_gene)

comp_res = compareCases(res_list = in_list,
                        excelOut = FALSE,
                        outFileName = 'comp_res')

}
