% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawingKaryo.R
\name{drawingKaryo}
\alias{drawingKaryo}
\title{Drawing the karyogram with integration sites.}
\usage{
drawingKaryo(hits, feature, organism = 'GRCh37', includeUndecided = FALSE, outPath = getwd(), 
             outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{hits}{a GR object. This object made from \code{makeInputObj} function.}

\item{feature}{a GR object. This object made from annotation function.}

\item{organism}{a character vector. This function serves 2 versions of organisms such as GRCh37, GRCh38 (Human).
Default is 'GRCh37'.}

\item{includeUndecided}{TRUE or FALSE. If user want to use undecided hits in analysis, enter TRUE.
Default is FALSE.}

\item{outPath}{a string vector. Type path to save a plot.}

\item{outFileName}{a character vector. This value used when saving the idegoram image file.}
}
\value{
Return the ideogram plot and object.
}
\description{
\preformatted{
This function draw an ideogram with the integration site data.
}
}
\examples{
data(blast_obj)
data(blast_gene)
drawingKaryo(hits = blast_obj, feature = blast_gene$Gene_data, outFileName = 'blast_res')

          
}
