% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integPatternByChr.R
\name{integPatternByChr}
\alias{integPatternByChr}
\title{Retroviral integration pattern on chromosomes}
\usage{
integPatternByChr(hits, ran_hits, excelOut = TRUE,
                  isExpList = FALSE, isRanList = FALSE,
                  outPath = getwd(),
                  outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{hits}{a GR object or list. This object made by \code{makeInputObj} function.}

\item{ran_hits}{a GR object or list. This object is output of \code{ranSetGenerator} function.}

\item{excelOut}{TRUE or FALSE. If user want to make excel file, enter TRUE. Default is TRUE.}

\item{isExpList}{TRUE or FALSE. If \item{hits} is the list of several experimental data,
enter TRUE. Default is FALSE.}

\item{isRanList}{TRUE or FALSE. If \item{ran_hits} is the list several random data,
enter TRUE. Default is FALSE.}

\item{outPath}{a string vector. Directory path of tab-deliminated hit files
generated by this function.}

\item{outFileName}{a character vector. Attached character to the result file name.}
}
\value{
Return a result list that is made up of table(s) about
        experimental/random data integration frequency on chromosomes.
}
\description{
Check retroviral integration pattern on chromosomes
}
\examples{
data(blast_obj)
chrCheck = integPatternByChr(hits = blast_obj)
 
}
