% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Make data files for RIPAT.}
\usage{
makeData(organism = 'GRCh37', dataType = 'gene')
}
\arguments{
\item{organism}{a single character. Two versions of organism such as GRCh37, GRCh38 (Human). Default is 'GRCh37'.}

\item{dataType}{a single character. Data type what user needs (gene, cpg, repeat and variant). Default is 'gene'.}
}
\value{
Database files are saved in the extdata directory of RIPAT.
}
\description{
Download datafiles for running RIPAT.
}
\examples{
makeData(organism = 'GRCh37')

}
