% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Make data files for RIPAT}
\usage{
makeData(organism = 'GRCh37')
}
\arguments{
\item{organism}{a single character. This function can run by 2 versions of organisms such as GRCh37, GRCh38 (Human). Default is 'GRCh37'.}
}
\description{
\preformatted{
RIPAT use specific format of data files for retroviral vector 
integration pattern analysis. This function made these data files from
well-known genomic databases easily. The data file location is extdata folder of RIPAT.
}
}
\examples{
#' 
makeData(organism = 'GRCh37')

}
