% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInputObj2.R
\name{makeInputObj2}
\alias{makeInputObj2}
\title{Make the retroviral vector integration site object.}
\usage{
makeInputObj2(inDir, id, mapTool = 'blast', 
              vectorPos = 'front', outPath = getwd(), 
              outFileName = paste0('RIPAT', round(unclass(Sys.time()))))
}
\arguments{
\item{inDir}{a string vector. Location of a directory that has a local alignment
result files. All alignment result files do not include 
any header and comment.}

\item{id}{a character vector. the specific words that can appoint 
alignment file names exclusively.}

\item{mapTool}{a character vector. Function serves two types of file
such as outputs from BLAST and BLAT. Default is 'blast'.
If you want to use BLAT result, use 'blat'.}

\item{vectorPos}{a character vector. Sets the position of vector on sequences.
Default value is 'front'. If the vector is located at the behind of sequence,
you can change it to 'behind'.}

\item{outPath}{a string vector. Directory path of tab-deliminated hit files
generated by this function.}

\item{outFileName}{a character vector. Attached character to the result file name.}
}
\value{
Return two types of outputs. Text file and R object.
        Available hit data from input is written in text file and
        generated as a list of GenomicRange(GR) format object.
}
\description{
Make an input object for annotation functions.
}
\examples{
blast_obj = makeInputObj2(inDir = system.file("scripts", package = "RIPAT"), id = 'BLASTn')
 
}
